Renamed namespace in Models folder to not include "Data". All changes are due to renaming this namespace.
