+++ ChessWPF/Models/Cells/ConstructorCell.cs
@@ -1,8 +1,8 @@
 ﻿using ChessWPF.Contracts.Pieces;
-using ChessWPF.Models.Data.Pieces;
+using ChessWPF.Models.Pieces;
 using System;
 
-namespace ChessWPF.Models.Data.Board
+namespace ChessWPF.Models.Cells
 {
     public sealed class ConstructorCell
     {
