+++ ChessWPF/Models/Options/GameOptions.cs
@@ -1,6 +1,6 @@
-﻿using ChessWPF.Models.Data.Enums;
+﻿using ChessWPF.Models.Enums;
 
-namespace ChessWPF.Models.Data.Options
+namespace ChessWPF.Models.Options
 {
     public class GameOptions
     {
