+++ ChessWPF/Models/Pieces/ConstructorMenuPiece.cs
@@ -1,5 +1,5 @@
 ﻿using ChessWPF.Contracts.Pieces;
-using ChessWPF.Models.Data.Pieces.Enums;
+using ChessWPF.Models.Pieces.Enums;
 
 namespace ChessWPF.Models.Pieces
 {
