+++ ChessWPF/Models/Pieces/Pawn.cs
@@ -1,7 +1,6 @@
-﻿using ChessWPF.Models.Data.Board;
-using ChessWPF.Models.Data.Pieces.Enums;
+﻿using ChessWPF.Models.Pieces.Enums;
 
-namespace ChessWPF.Models.Data.Pieces
+namespace ChessWPF.Models.Pieces
 {
     public sealed class Pawn : Piece
     {
