+++ ChessWPF/Models/Pieces/Piece.cs
@@ -1,8 +1,8 @@
-﻿using ChessWPF.Models.Data.Board;
-using ChessWPF.Models.Data.Pieces.Enums;
+﻿using ChessWPF.Models.Cells;
+using ChessWPF.Models.Pieces.Enums;
 using System.Collections.Generic;
 
-namespace ChessWPF.Models.Data.Pieces
+namespace ChessWPF.Models.Pieces
 {
     public abstract class Piece
     {
